New command create-dev-settings. Clean up main script. Fix wrong path for Windows version.
