+++ .gitignore
@@ -7,10 +7,12 @@
 .virtualenv/*
 .venv/*
 
-# Development settings and database
+# Development user data (settings, database, media, search index)
 settings.py
-database.sqlite
 !tests/settings.py
+database.sqlite
+media/*
+whoosh_index/*
 
 # Package building
 docs/_build/*
@@ -21,4 +23,4 @@ dist/*
 
 # Unit test and coverage reports
 .coverage
-htmlcov
+htmlcov/*
