+++ openslides/utils/settings.py.tpl
@@ -1,19 +1,19 @@
 # -*- coding: utf-8 -*-
 #
-# Settings file for OpenSlides' tests
+# Settings file for OpenSlides
 #
 
-
-from openslides.global_settings import *  # noqa
+%(import_function)s
+from openslides.global_settings import *
 
 # Use 'DEBUG = True' to get more details for server errors. Default for releases: False
-DEBUG = True
+DEBUG = %(debug)s
 TEMPLATE_DEBUG = DEBUG
 
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': '',
+        'NAME': %(database_path_value)s,
         'USER': '',
         'PASSWORD': '',
         'HOST': '',
@@ -25,19 +25,18 @@ DATABASES = {
 TIME_ZONE = 'Europe/Berlin'
 
 # Make this unique, and don't share it with anybody.
-SECRET_KEY = 'secred'
+SECRET_KEY = %(secret_key)r
 
 # Add OpenSlides plugins to this list (see example entry in comment)
 INSTALLED_PLUGINS = (
-    'tests.person_api',
+#    'pluginname',
 )
 
 INSTALLED_APPS += INSTALLED_PLUGINS
 
 # Absolute path to the directory that holds media.
 # Example: "/home/media/media.lawrence.com/"
-MEDIA_ROOT = os.path.realpath(os.path.dirname(__file__))
+MEDIA_ROOT = %(media_path_value)s
 
 # Path to Whoosh search index
-# Use RAM storage
-HAYSTACK_CONNECTIONS['default']['STORAGE'] = 'ram'
+HAYSTACK_CONNECTIONS['default']['PATH'] = %(whoosh_index_path_value)s
